requirements.txt
click==8.1.7 
structlog==24.4.0 
pydantic==2.9.2 
requests==2.32.3 
pyyaml==6.0.2
rich==13.9.2